Allow querying the contents of settings.yml (and settings_user!) from the API
