+++ conan/api/subapi/config.py
@@ -1,14 +1,17 @@
 import os
 import platform
 import textwrap
-
+import yaml
 from jinja2 import Environment, FileSystemLoader
 
 from conan import conan_version
+from conans.client.conf import default_settings_yml
 from conan.internal.api import detect_api
 from conan.internal.cache.home_paths import HomePaths
 from conan.internal.conan_app import ConanApp
-from conans.model.conf import ConfDefinition
+from conans.errors import ConanException
+from conans.model.conf import ConfDefinition, BUILT_IN_CONFS
+from conans.model.settings import Settings
 from conans.util.files import load, save
 
 
@@ -68,3 +71,47 @@ class ConfigAPI:
                     """)
                 save(global_conf_path, default_global_conf)
         return self._new_config
+
+    @staticmethod
+    def builtin_confs():
+        return BUILT_IN_CONFS
+
+    @property
+    def settings_yml(self):
+        """Returns {setting: [value, ...]} defining all the possible
+                   settings without values"""
+        _home_paths = HomePaths(self.conan_api.cache_folder)
+        settings_path = _home_paths.settings_path
+        if not os.path.exists(settings_path):
+            save(settings_path, default_settings_yml)
+            save(settings_path + ".orig", default_settings_yml)  # stores a copy, to check migrations
+
+        def _load_settings(path):
+            try:
+                return yaml.safe_load(load(path)) or {}
+            except yaml.YAMLError as ye:
+                raise ConanException("Invalid settings.yml format: {}".format(ye))
+
+        settings = _load_settings(settings_path)
+        user_settings_file = _home_paths.settings_path_user
+        if os.path.exists(user_settings_file):
+            settings_user = _load_settings(user_settings_file)
+
+            def appending_recursive_dict_update(d, u):
+                # Not the same behavior as conandata_update, because this append lists
+                for k, v in u.items():
+                    if isinstance(v, list):
+                        current = d.get(k) or []
+                        d[k] = current + [value for value in v if value not in current]
+                    elif isinstance(v, dict):
+                        current = d.get(k) or {}
+                        if isinstance(current, list):  # convert to dict lists
+                            current = {k: None for k in current}
+                        d[k] = appending_recursive_dict_update(current, v)
+                    else:
+                        d[k] = v
+                return d
+
+            appending_recursive_dict_update(settings, settings_user)
+
+        return Settings(settings)
